requirements.txt
torch
soundfile
onnxruntime
pandas
ffmpeg-python
matplotlib
